changed int to size_t because unsigned

+++ BTreeNode.h
@@ -10,31 +10,28 @@ class BTreeNode
 public:
 	vector<Book> keys;
 	std::vector<std::unique_ptr<BTreeNode>> children;
-	// minimum degree
 	int t;
 	bool leaf;
-	
+
 	BTreeNode(int _t, bool leaf);
 
-
 	void traverse();
-	BTreeNode* search(const string& isbn);
-	void insertNonFull(const Book& book);
-	void splitChild(int i, std::unique_ptr<BTreeNode>& y);
-	int findKey(const string& isbn);
-	void remove(const string& isbn);
-	void removeFromLeaf(int idx);
-	void removeFromNonLeaf(int idx);
-	Book getPred(int idx);
-	Book getSucc(int idx);
-	void fill(int idx);
-	void borrowFromPrev(int idx);
-	void borrowFromNext(int idx);
-	void merge(int idx);
+	BTreeNode *search(const string &isbn);
+	void insertNonFull(const Book &book);
+	void splitChild(int i, std::unique_ptr<BTreeNode> &y);
+	size_t findKey(const string &isbn);
+	void remove(const string &isbn);
+	void removeFromLeaf(size_t idx);
+	void removeFromNonLeaf(size_t idx);
+	Book getPred(size_t idx);		
+	Book getSucc(size_t idx);		
+	void fill(size_t idx);			
+	void borrowFromPrev(size_t idx);
+	void borrowFromNext(size_t idx);
+	void merge(size_t idx);			
 	void printNode();
 
 	friend class BTree;
 
 	~BTreeNode() = default;
-
 };
